+++ src/terminal_sandbox.mli
@@ -1,7 +1,9 @@
-type t
+type t = Vscode.Terminal.t
 
-val create : Sandbox.t -> t
+val create : ?name:string -> ?command:Cmd.spawn -> Sandbox.t -> t
 
 val dispose : t -> unit
 
 val show : t -> unit
+
+val send : t -> string -> unit