+++ vscode/vscode/vscode.ml
@@ -276,6 +276,10 @@ end
 module Selection = struct
   include Class.Extend (Range) ()
 
+  include Range
+
+  type t = private (* class extends *) Range.t [@@js]
+
   val anchor : t -> Position.t [@@js.get]
 
   val active : t -> Position.t [@@js.get]