+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## Unreleased
+## 1.7.0
 
 - Fixed an issue when uninstalled Opam packages still appear in the `roots`
   field of Opam's `switch-state` file (#517)